+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Postman Runtime Changelog
 
+#### 2.2.0 (July 19, 2016)
+* Fixed a bug which caused `done` to be called twice if it threw an error
+* Added an option to abort a run on test failures (as well as errors)
+
 #### 2.1.1 (July 13, 2016)
 * Updated to postman-collection@0.4.1
 * Fixed a bug that caused the Runtime to crash when used with stopOnError and with multiple iterations